requirements.txt
beautifulsoup4
evaluate
llama-cpp-python
matplotlib
nltk
numpy
pandas
pydub
requests
scikit-learn
spacy
SpeechRecognition
sentencepiece
torch
torchmetrics
torchvision
transformers
wordcloud
